server/requirements.txt
fastapi[standard]
uvicorn==0.32.0
sqlalchemy==2.0.36
sqlalchemy_utils==0.41.2
asyncpg==0.30.0
pydantic_settings==2.6.1
cryptography==43.0.3
pyjwt==2.10.0
passlib==1.7.4
python-multipart==0.0.17
alembic==1.16.1
greenlet==3.2.2
boto3==1.38.28
celery==5.5.3
redis==6.2.0
slowapi==0.1.9
firebase-admin==6.5.0